more adoption of sklearn api
